PrintFloatVariable modified double to float

+++ 0x00-java-hello_world/PrintFloatVariable.java
@@ -1,6 +1,6 @@
 public class PrintFloatVariable {
   public static void main(String[] args) {
-    double pi = 3.14159;
+    float pi = 3.14159f;
     System.out.printf("Pi Value: %.3f\n", pi);
   }
 }
